# webhook-event-publisher-api/oas-spec/src/main/resources/openapi/event-validation-api.yml
openapi: 3.0.3
info:
  title: Event Validation API
  version: 0.1.0
  description: |
    API for validating webhook events, retrieving schema metadata, and publishing to MSK ingress topics.
servers:
  - url: https://api.mycorp.example
    description: Production
  - url: https://api.sandbox.mycorp.example
    description: Sandbox
paths:
  /events/{eventName}:
    post:
      summary: Validate and enqueue an event payload
      description: >-
        Validates an inbound payload against the configured schema and, if successful,
        publishes it to the matching MSK ingress topic.
      operationId: publishEvent
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
          description: Logical event name (e.g., CustomerUpdated)
        - name: producer-domain
          in: header
          required: true
          schema:
            type: string
          description: Producer domain used for routing and topic derivation
        - name: schema-id
          in: header
          required: true
          schema:
            type: string
          description: Identifier of the schema stored in DynamoDB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPayload'
      responses:
        '202':
          description: Payload accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate or conflicting idempotency key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Downstream Kafka unavailable or throttled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /schemas/{domain}/{event}/{version}:
    get:
      summary: Retrieve schema metadata for a specific event version
      operationId: fetchSchema
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
        - name: event
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schema metadata located
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaMetadata'
        '404':
          description: Schema not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    EventPayload:
      type: object
      required:
        - eventId
        - occurredAt
        - data
      properties:
        eventId:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
        correlationId:
          type: string
        schemaVersion:
          type: string
        data:
          type: object
          additionalProperties: true
    AckResponse:
      type: object
      required:
        - status
        - eventId
      properties:
        status:
          type: string
          example: ACCEPTED
        eventId:
          type: string
          format: uuid
        enqueueTimestamp:
          type: string
          format: date-time
    SchemaMetadata:
      type: object
      required:
        - schemaId
        - format
        - transformFlag
      properties:
        schemaId:
          type: string
        format:
          type: string
          enum: [JSON, XML, SOAP, AVRO]
        transformFlag:
          type: string
          enum: [Y, N]
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        eventName:
          type: string
        producerDomain:
          type: string
        schemaDefinition:
          type: string
          description: Underlying JSON Schema, XSD, or Avro definition
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        details:
          type: array
          items:
            type: string